New Database Structures and Connectionless Vagrant Running
Added first drafts for schemas and population data regarding event codes,
reference tables and pulses.  Definitely not set in stone.

Added the ability to run "Vagrant Up" without any connection whatsoever.
This means there are no more apt-get updates, downloads, etc.
The downside to this is that it requires local copies of debian
packages, the vagrant box and the docker image used.  The packages
will have to be updated when the image is updated, as some depend
on the specific versions contained in the docker image.

The easy trick to this is after creating the new docker image, run "apt-get
install" for whatever packages you need for whichever container you need,
grab the list of new packages installed, and then aptitude download them all
into whichever deb folder corresponds to the container (for example,
the postgres container has a folder "postgres-deb" for its packages).
For the pip installs, pretty much the same except there is a
bundle creation script to create the pip bundle from the requirements.txt.

Created "download_packages.sh" scripts in each of the package folders,
however I am having issues running them in a way that downloads the correct
packages for the docker version rather than packages for my OS.

Keep in mind that the packages, vagrant box and docker image will NOT be
subversioned.  They are much too large, for example, the vagrant box is
400 - 600 MB right now.

Modified the .gitignore to ignore all the packages in the container
package folders, as well as the vagrant box and docker image.

Added a new page to view pulses.  For now, choice is just by the pulse
id, not enough information about the requirements.  Uses its own
"pulse_dygraph.html", which is very similar to the "instrument_dygraph.html"
but has enough different to prevent cohesive generalizing, hence the split
from "dygraph.html" into two specific files.