Not Working. Event Manager mostly clean/functional, agent is not.
Will not run right out of the repository.  Event managers/agent are set
to run localhost rather than within vagrant right now.  My working
environment was 2 copies of the event manager folder, one with a config.yml
that said it was the central facility and one that did not, with the central
facility manager being hosted on localhost port 5001 instead of 5000.

Working
+Agent request event code / propagates up chain / created if not exists
passed back down chain and saved
+Site id request goes up and down the chain, with the Site agent saving
all of the information if it did not already have it
+Same thing for instruments
+If event code passed to managers is not a "special" code (meaning its
own processing chain) the event is saved in both the Site and CF Managers
+One agent plugin pulls the status of ena-kazr and stores the attributes as
individual events(proof of concept/functionality, will eventually be one
large event saved in its specialized table).

Those are the currently recalled highlights.  Next item to do is to have a
special event code for the PAFClient saves in the database (and saving func)