Not working. Event manager now runs and communicates with agent within vagrant.
For now, runs the agent and the site event manager, and tries to communicate
with a central facility event manager on whatever machine is running the
vagrant box.  I've been using a copy of the event manager with a copy of
the database schema and data files.  In order to run correctly, the event code
table of the central facility database must be
1 "event_code_request"
2 "site_id_request"
3 "instrument_id_request"
4 "prosensing_paf"

Next step is to set this up for multi-machine running in vagrant
One machine for the site event manager, one for the central facility.