Set up the beginnings of a multi machine environment.
Looks like so far, to get this running on one machine will require two copies
of the warno vagrant system, one to run as central and one to run as site.
This is because some of the environment and config files would be shared
if they were located in the same directory.

------------------------------------
To get the system running:

Copy the system into two directories, one for "central" and one for "site".

For the central directory, make the following modifications:
+db_up db address to 192.168.50.100
+db_save db address to 192.168.50.100
+eventmanager/src/database utility.py address to 192.168.50.100 in three
 places: db_address at the top and in the two engine calls in the body
+change eventmanager config.yml setup:central: from 0 to 1

In the central directory, do "vagrant up central"
It will load a db_dump file if present, if not, you should run
eventmanager/src/database populate_db.py to load in basic data
(Note: Not necessary to do this in the "site" directory, as it will
pull the data it needs from "central")

After central is running smoothly, do "vagrant up site".
This should start up the event manager and the agent, communicating and
saving to central.
----------------------------------

May want to remove the prosensing_paf.data entries before running or
there will be a large time discrepency between the demo data and the
current data.

The first time you vagrant up site may fail because the agent may start
before the database is initialized. Looking into finding how to delay agent
from starting until it's ready.

Currently, the central agent (which shouldn't start up) starts up and
crashes when it doesn't get the expected data.  Should look up a better
way to stop the agent from starting than an expected failure.