Added support for vagrant to communicate with test servers.
Modified machines to route port 8080 on host to port 80 on vm.  Apparently
it is very insecure to route 80 on host to 80 on vm, so vagrant does
not allow this.

Modified db_save.sh and db_up.sh to be more fault tolerant, and if there
is no database dump file to load from, to initialize a new database.