Site appearance overhaul, graphing by date/time range functionality.
Updated site appearance.
Adding instrument data graphs now accepts a specified start datetime and
end datetime to specify the range of data to be retrieved.  Datetimes
are in UTC.

Minor edit to submitting a log, chrome had a default picker
for the log date, but the return from it was considered invalid for our
code.  Disabled the chrome-specific picker.