[NOT WORKING] Progress toward using data volumes for postgres.
Fixed a typo in RadarStatusMonitor.py that prevented updating instrument type.

Data Volumes for Postgres not working correctly, finally tracked it down to
an issue where the database is initialized properly in the dockerfile, but is
overwritten when it takes the "volumes_from" the data container.  Fixing looks
to be complicated.