Reverted back to single machine configuration, data container for database.
db_up.sh and db_save.sh are being phased out, still in repo for reference.

Added a new data container constructed similar to the postgres container.
Allows for persistent storage between container and vm restarts.
However, if the vm is destroyed, this data volume will be as well.
Regular backup system recommended to put in place.

If two vms on one machine are still necessary:
For the second machine:
   Vagrantfile:
      Change the ip to 192.168.50.99
      Change the vm hostname and virtualbox name away from "warno"
   event_manager/src/database/utility.py:
      Change 192.168.50.100 to 192.168.50.100 in three places
   event_manager/src/warno_event_manager.py:
      Change database address from 192.168.50.100 to 192.168.50.100
   user_portal/src/RadarStatusMonitor.py:
      Change database address from 192.168.50.100 to 192.168.50.99